requirements.txt
Django
python-dateutil
fontawesomefree==6.6.0